added getFoodNames and getReviews(rest_name) methods

+++ src/com/ruchi/hibernate/model/DAO/ReviewDao.java
@@ -1,7 +1,6 @@
 package com.ruchi.hibernate.model.DAO;
 
 import javax.persistence.*;
-import java.util.Date;
 
 /**
  * Created by Thamayanthy on 12/22/2014.
@@ -26,37 +25,35 @@ public class ReviewDao {
     @Column(name = "rating", nullable = true, unique = false, length = 50)
     private Float rating;
 
-    public String getReview_id() {
-        return review_id;
-    }
+	public String getReview_id() {
+		return review_id;
+	}
 
-    public void setReview_id(String review_id) {
-        this.review_id = review_id;
-    }
+	public void setReview_id(String review_id) {
+		this.review_id = review_id;
+	}
 
-    public String getRest_id() {
-        return rest_id;
-    }
+	public String getRest_id() {
+		return rest_id;
+	}
 
-    public void setRest_id(String rest_id) {
-        this.rest_id = rest_id;
-    }
+	public void setRest_id(String rest_id) {
+		this.rest_id = rest_id;
+	}
 
-    public String getReview() {
-        return review;
-    }
+	public String getReview() {
+		return review;
+	}
 
-    public void setReview(String review) {
-        this.review = review;
-    }
+	public void setReview(String review) {
+		this.review = review;
+	}
 
-	public float getRating() {
-		return rating.floatValue();
+	public Float getRating() {
+		return rating;
 	}
 
 	public void setRating(Float rating) {
 		this.rating = rating;
-	}
-
-  
+	}  
 }
